Populating the world. Two new pirates added.

+++ Turn based game/Assets/Scripts/GameData.cs
@@ -6,7 +6,8 @@ public class GameData
 
     public int[] charactersInvolvedInCombat;
     public int[] charsHealth;
-    public int[] charsDamageBoost;
+
+    public float[] charsDamageBoost;
 
     public float[,] charsPositions;
